+++ lib/elixir/lib/hash_dict.ex
@@ -129,17 +129,17 @@ defmodule HashDict do
     index = key_mask(hash)
     case elem(node, index) do
       [] ->
-        {set_elem(node, index, [key|value]), 1}
+        {put_elem(node, index, [key|value]), 1}
       [^key|_] ->
-        {set_elem(node, index, [key|value]), 0}
+        {put_elem(node, index, [key|value]), 0}
       [k|v] ->
-        n = set_elem(@node_template, key_mask(key_shift(hash)), [key|value])
-        {set_elem(node, index, {k, v, n}), 1}
+        n = put_elem(@node_template, key_mask(key_shift(hash)), [key|value])
+        {put_elem(node, index, {k, v, n}), 1}
       {^key, _, n} ->
-        {set_elem(node, index, {key, value, n}), 0}
+        {put_elem(node, index, {key, value, n}), 0}
       {k, v, n} ->
         {n, counter} = do_put(n, key, value, key_shift(hash))
-        {set_elem(node, index, {k, v, n}), counter}
+        {put_elem(node, index, {k, v, n}), counter}
     end
   end
 
@@ -147,17 +147,17 @@ defmodule HashDict do
     index = key_mask(hash)
     case elem(node, index) do
       [] ->
-        {set_elem(node, index, [key|initial.()]), 1}
+        {put_elem(node, index, [key|initial.()]), 1}
       [^key|value] ->
-        {set_elem(node, index, [key|fun.(value)]), 0}
+        {put_elem(node, index, [key|fun.(value)]), 0}
       [k|v] ->
-        n = set_elem(@node_template, key_mask(key_shift(hash)), [key|initial.()])
-        {set_elem(node, index, {k, v, n}), 1}
+        n = put_elem(@node_template, key_mask(key_shift(hash)), [key|initial.()])
+        {put_elem(node, index, {k, v, n}), 1}
       {^key, value, n} ->
-        {set_elem(node, index, {key, fun.(value), n}), 0}
+        {put_elem(node, index, {key, fun.(value), n}), 0}
       {k, v, n} ->
         {n, counter} = do_update(n, key, initial, fun, key_shift(hash))
-        {set_elem(node, index, {k, v, n}), counter}
+        {put_elem(node, index, {k, v, n}), counter}
     end
   end
 
@@ -167,17 +167,17 @@ defmodule HashDict do
       [] ->
         :error
       [^key|value] ->
-        {set_elem(node, index, []), value}
+        {put_elem(node, index, []), value}
       [_|_] ->
         :error
       {^key, value, n} ->
-        {set_elem(node, index, do_compact_node(n)), value}
+        {put_elem(node, index, do_compact_node(n)), value}
       {k, v, n} ->
         case do_delete(n, key, key_shift(hash)) do
           {@node_template, value} ->
-            {set_elem(node, index, [k|v]), value}
+            {put_elem(node, index, [k|v]), value}
           {n, value} ->
-            {set_elem(node, index, {k, v, n}), value}
+            {put_elem(node, index, {k, v, n}), value}
           :error ->
             :error
         end
@@ -188,12 +188,12 @@ defmodule HashDict do
     defp do_compact_node(node) when elem(node, unquote(index)) != [] do
       case elem(node, unquote(index)) do
         [k|v] ->
-          case set_elem(node, unquote(index), []) do
+          case put_elem(node, unquote(index), []) do
             @node_template -> [k|v]
             n -> {k, v, n}
           end
         {k, v, n} ->
-          {k, v, set_elem(node, unquote(index), do_compact_node(n))}
+          {k, v, put_elem(node, unquote(index), do_compact_node(n))}
       end
     end
   end